Added client tags and fixed IClientData
Added functionality to tag clients for messages to be sent to distinct groups. This is useful for a variety of reasons:
Private groups/messages,
Region based updates (only send information to clients that need it instead of sending all information to everyone, which is bad for obvious reasons)

In creating this functionality, I discovered that my implementation of ClientData was just flat out wrong. Somehow in my brain I could have a static instance of ClientData floating around which would be instantiated for each new client, but that was just wrong and every connected client had the same ClientData, which is certainly not intended. Doing some research, I found that to achieve what I wanted I would have to use reflection. Definitely a learning experience, but damn reflection is useful. I'll definitely have to improve upon this on the client side as well, not to mention a lot of code that probably doesn't need to be there, but I'm already overdue on a commit and this works how I want it to.